Move all ui components to the related package.
