+++ app/src/main/java/ru/sandbox/androidacademyapp/ui/MoviesItemDecoration.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp
+package ru.sandbox.androidacademyapp.ui
 
 import android.graphics.Rect
 import android.view.View
