+++ app/src/main/java/ru/sandbox/androidacademyapp/ui/MoviesViewModelFactory.kt
@@ -1,4 +1,4 @@
-package ru.sandbox.androidacademyapp
+package ru.sandbox.androidacademyapp.ui
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
